refactor(auth): rename project
