+++ README.md
@@ -1,7 +1,7 @@
-![CI](https://github.com/Liftric/auth/workflows/CI/badge.svg) 
-![maven-central](https://img.shields.io/maven-central/v/com.liftric/auth?label=Maven%20Central) 
-![OSS Sonatype (Releases)](https://img.shields.io/nexus/r/com.liftric/auth?label=Sonatype%20OSSRH%20%28Releases%29&server=https%3A%2F%2Fs01.oss.sonatype.org)
-![npm (scoped)](https://img.shields.io/npm/v/@liftric/auth)
+![CI](https://github.com/Liftric/cognito-idp/workflows/CI/badge.svg) 
+![maven-central](https://img.shields.io/maven-central/v/com.liftric/cognito-idp?label=Maven%20Central) 
+![OSS Sonatype (Releases)](https://img.shields.io/nexus/r/com.liftric/cognito-idp?label=Sonatype%20OSSRH%20%28Releases%29&server=https%3A%2F%2Fs01.oss.sonatype.org)
+![npm (scoped)](https://img.shields.io/npm/v/@liftric/cognito-idp)
 
 # Cognito-idp
