+++ cognito-idp/src/androidTest/kotlin/com/liftric/cognito-idp/IdentityProviderClientTests.kt
