+++ cognito-idp/src/commonMain/kotlin/com/liftric/cognito-idp/Configuration.kt
