+++ cognito-idp/src/commonMain/kotlin/com/liftric/cognito-idp/core/Exceptions.kt
