+++ cognito-idp/src/main/AndroidManifest.xml
